Collaborative Development of a Second-Hand Marketplace-CompositionService-main/requirements.txt
fastapi==0.75.1
httpx==0.23.0
uvicorn==0.18.2
python-dotenv==0.20.0
